Measure buckets with timestamps
Waiting for buckets of bits to go through with a fixed-count loop of usleep() calls created a lot of variability in how many bits were in each measurement. Now we use a hardware timer provided by the device to measure cumulative time accurately.

Resulting timing variation is now greatly reduced. The overhead of surrounding statements is still present, along with the uncertain duration of the single last usleep() call.