Fix segmentation fault at first dummy frame
Fixed a buffer overrun on dummy_frame in send_dummy_frame due to a constant byte count that had not been updated. Fixes issue #21.

Added/enhanced a number of debug printf calls. Some messages are now tagged with "time" in the value of last_frame_time.tv_nsec. This not really time, though, because it is maintained by repeated adding exactly 40ms per frame. Unintended delays (observed on spectrum analyzer but not yet understood) are not reflected in this tag.

Changed type of elapsed_ms in check_hang_timer from uint32_t to int32_t. This is important because elapsed_ms is computed as the difference between two nanosecond timers, and this value can be negative. In particular, it's always negative after the transmission of a preamble frame, since that transmission adds 40ms to last_frame_time. This caused the transmission of dummy frames to begin prematurely, which led to early seg faults. Fixes issue #20.

Note that when this program on the target is being run within an ssh command on the host, the segmentation fault error message is not visible. Before it has time to show up on the host, the ssh connection is torn down. Thus instead of "Segmentation fault" we see just "Connection to <target ip> closed." In order to see the "Segmentation fault" message we had to ssh into the target and then run this program from the root shell on the target.