Feed modulator 271 bytes/frame
We've been sending the logical data to the modulator: 12 bytes of header, 122 bytes of payload, total of 134 bytes. That would be appropriate if the modulator took over added the FEC and the frame sync word, but at the moment the modulator doesn't do any of that.

Now this program creates each frame type appropriately and always sends 271 bytes of data to the modulator for each frame. For a normal frame, that's three bytes of frame sync word, 24 bytes encoding the 12 bytes of header at rate one-half, and 244 bytes encoding the 122 bytes of payload at rate one-half. For the preamble it's just 271 bytes of the pattern. For the postamble, it's three bytes of frame sync word, 24 bytes of encoded header, and 244 bytes of the pattern. For dummy frames, it's like a normal frame except that the payload is all zeroes.

The frame sync word is real, selected by concatenating the 11-bit Barker code and the 13-bit Barker code, in that order.

The FEC is fake for now. We just repeat the header to simulate encoding it at rate one-half, and likewise for the payload. Replace functions encode_ovp_header and encode_ovp_payload with real FEC to make this real.

We've also been running the modulator with TX_DATA_WIDTH set to 32, without me being aware of that. That means the modulator was accepting 4 bytes of data with every AXI-S transfer (32 bits in this case). But we were only packing a single data byte into the transfer, and oddly enough we were packing it in twice. Since the number of bytes we are sending each time is odd (271), we have to set the TX_DATA_WIDTH to 8 (which is the default) and put a single byte in the right place in each transfer. This is now done.

Fixed the semantics of the axis_xfer_count delta logging.

Also found a fixed some typos and comments.