Deterministic timeline reorg
THIS COMMIT COMPILES CLEANLY BUT IS NOT TESTED.

Reorganized transmit frame handling to gain full control of
the timeline. We use the hardware timestamp at microsecond
resolution for precision, and arrange to always decide about
what kind of frame (normal, dummy, or postamble) to send at
halfway through the next frame slot. This allows plus or minus
half a frame of jitter in the arrival times of the encapsulated
frames.

There are now two additional threads. Up to now, we had one
"listener" thread for receiving UDP-encapsulated frames from
the network, and that thread also handled ushering those frames
through the pipeline. The main thread handled frame underrun
(including end of transmission) and periodic statistics
reporting, all asynchronously to the main timeline. Now, the
function of pushing the frame data toward the modulator is
removed from the listener thread. A new thread, the timeline
manager, wakes up at halfway through the frame slot and
decides whether an encapsulated frame has been received.
If so, it pushes it toward the modulator in plenty of time
to avoid underrunning the IIO buffers. If not, it instead
pushes a dummy frame as part of the "hang time", or else
a postamble frame if the hang time is completed. Another
new thread handles emitting the periodic status report,
mostly for symmetry and to clean up the main loop in the
main thread, which is now just repeated sleeps until the
program is stopped.

All threads do their work inside a shared mutex.

The two new threads shut down operations between transmission
sessions.

In this new timeline, we're also able to detect and handle
the case where more than one encapsulated frame arrives within
the window of a single frame slot. We count the occurrences
of this event (which we report as "untimely frames") and then
choose simply to accept only the last arriving frame and
discard the others.

All frames (except preambles) are now built in a global buffer
called modulator_frame_buffer. This makes it easy for dummy
frames, and then finally the postamble frame, to be derived
from the last transmitted normal frame. Along with this
change, there's some optimization in that we only compute
the dummy and postamble payloads once.

There's now a distinction between the normal end of a frame
(which lets the postamble go out and shuts down cleanly) and
an aborted end of a frame (which just turns everything off,
in order to exit the program).

send_ovp_frame_to_msk() has been split up into
load_ovp_frame_into_txbuf() and push_tx_buf_to_msk()
because these two operations now take place at different times.

The timeline is fixed as soon as the first encapsulated frame
is received. The decision times are spaced 40ms apart, with
the reference time being 60ms after the arrival of the first
frame. There is no tracking or drift, unless there's a
disagreement between the encapsulated frame source and this
program's hardware timestamp clock about how long 40ms is.

No use is made of the Linux CLOCK_MONOTONIC source of time.

The wait after sending the postamble before turning off the
transmitter is increased from 50ms to 65ms. It might reasonably
take 60ms to get through the timeline and buffers, plus 5ms
of margin. This does not (and cannot) guarantee that the
postamble is completely transmitted before the transmitter
is shut off, but it usually should.

Before this, postamble frames used a special authentication
tag in the frame header. This was supposed to make it easier
for the receiver to detect the postamble, but it opens a
security hole: anyone could transmit a postamble and the
satellite (or groundsat, etc.) would be unable to detect it.
Now the postamble and all the dummy frames use the authentication
tag from the most recent normal frame. This is not completely
secure in that an attacker close enough to a transmitting
station could eavesdrop on the uplink and replay the header,
it's still better than a fixed authentication tag.

Removed some unused flexibility in frame buffer sizes.

Tightened up line spacing, fixed some spelling errors,
inconsistent indentation, etc.