Updates for testing scenarios; OVP_FRAME_MODE frozen
Changed the fake stream transmission data pattern to better see what
is going on. Now we transmit 32 copies of a one-byte code that can be
recognized even with the wrong bit alignment, and then fill the rest
of the frame with 0xaa or 0x55, on alternate frames.

Added code to dump the fake stream frames to the screen.

In STREAMING mode, changed receive operations to depend on defining
RX_ACTIVE. So by default, streaming is transmit-only now.

Added, then commented out, code to pace the frame transmissions to
occur at 40ms intervals.

NOTE: OVP_FRAME_MODE in opulent_voice.c is frozen and no longer
maintained. Code from opulent_voice.c that is needed for
OVP_FRAME_MODE has been copied into a new file, dialogus.c,
which will be committed when ready.