Support 134-byte logical frame interface with modem
Added runtime switches software_tx_processing and software_rx_processing. Together, these switches control whether the randomization, FEC, and interleaving take place in software or in PL hardware. This includes changing the IIO buffer sizes to match the size of frames as they are created and consumed by the software. We further assume that if the hardware is handling frame processing, it also handles generation of the preamble.

All transmit frame processing (payload, dummy, and postamble) is routed through logical_frame_buffer (134 bytes). Then if software_tx_processing is enabled, the processing steps occur at decision time, transferring the frame to modulator_frame_buffer just in time for it to be copied into txbuf and pushed to IIO. With software_tx_processing turned off, it just copies the logical_frame_buffer into txbuf and pushes that to IIO, leaving the rest of processing to the modem.

Enabled hardware-generated preambles when software_tx_processing is false, because in that case we have no way to send a pattern of data directly to the modulator from software. We could and probably should use hardware-generated preambles all the time, but not today.

Build dummy frames and postamble frames in the logical_frame_buffer. That enables the dummy and postamble frames to follow the same path as the payload frames. The build process for dummy frames is just a memset to zero out the payload part of the frame, keeping the frame header from the preceding frame. The build process for postamble frames is similar, but uses a precomputed pattern and a memcpy in place of the memset. There's a Python snippet embedded in the code that shows how to produce the precomputed pattern.

We don't pass around frame lengths nearly as much as before. Instead, most of the time we can infer the frame length of 134 or 268 bytes. This eliminates a lot of useless error checking and redundant ways to register a frame error.

Because of changes already made in the HDL development, FEC is no longer separate for frame header and frame payload. One big FEC over the whole frame gives slightly better performance, and doesn't seem to cause any problems.

The frame handling steps have been simplified a little and made more consistent, and the functions given names that seem less confusing.

Reduce default output from debug thread. Add dump of bytes going into txbuf.

Removed the "success rate" statistic from the periodic report, since there's no longer a way for frame processing to fail in that way.

Adopted the term "decapsulated" to describe frames that arrived encapsulated but are no longer encapsulated. Adopted the term "accept" for early work on a decapsulated frame, distinct from "process" which I'm trying to use just for the (randomization, FEC, interleaving) processing that may or may not be done by the software. Later, a frame is said to be "readied for push" if it has been put into logical_frame_buffer. Processing, if done in software, copies it into modulator_frame_buffer. The actual iio_buffer_push is done from whichever buffer is appropriate.

Fixes #29