Separate IIO timeouts for transmit and receive
Use different IIO timeouts for transmit and receive. This setting
is associated with the context, according to the function naming.
Transmit uses the initially created context, ctx, where the IIO
timeout is set to 1 second (which is also the default). After we
clone the context (which is supposed to help with thread safety),
we change the timeout in the cloned rx_ctx to be infinite. Before,
we had used infinite timeouts for both, but we ran into a case
where transmit calls to iio_buffer_push were hanging the whole
system. Better that they time out, though in this case the problem
did not resolve itself.