Calibrate timestamp clock for different targets
Use the value of BogoMIPS reported in /proc/cpuinfo. That value (in MHz) is
half the CPU clock, and the conversion factor between global timer ticks and
seconds is half of that. So on ADALM PLUTO, with CPU clock of 666.666... MHz,
the factor is 166,666,666.67, whereas on LibreSDR with base clocking, the CPU
clock of 750 MHz yields a factor of 187,500,000.00.